requirements.txt
numpy
pandas
opencv-python
scikit-learn
scikit-image
matplotlib
pillow
psycopg2-binary
face-recognition
dlib